refactor: dto request와 response 패키지 분리 및 클래스명 수정, 중복 확인에 사용하는 jpa 메서드 변경
