+++ src/main/java/com/WearWeather/wear/user/controller/UserController.java
@@ -2,8 +2,8 @@ package com.WearWeather.wear.user.controller;
 
 import com.WearWeather.wear.global.common.ResponseMessage;
 import com.WearWeather.wear.global.common.dto.ResponseCommonDTO;
-import com.WearWeather.wear.user.dto.RequestRegisterUserDTO;
-import com.WearWeather.wear.user.dto.ResponseDuplicateCheckDTO;
+import com.WearWeather.wear.user.dto.request.RegisterUserRequest;
+import com.WearWeather.wear.user.dto.response.NicknameDuplicateCheckResponse;
 import com.WearWeather.wear.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -20,17 +20,17 @@ public class UserController {
     private final UserService userService;
 
     @PostMapping("/register")
-    public ResponseEntity<ResponseCommonDTO> signup(@Valid @RequestBody RequestRegisterUserDTO registerUserDTO){
+    public ResponseEntity<ResponseCommonDTO> signup(@Valid @RequestBody RegisterUserRequest registerUserDTO){
 
         userService.registerUser(registerUserDTO);
         return ResponseEntity.ok(new ResponseCommonDTO(true, "User registered successfully."));
     }
 
     @GetMapping("/nickname-check/{nickname}")
-    public ResponseEntity<ResponseDuplicateCheckDTO> checkDuplicateNickname(@PathVariable("nickname") String nickname){
+    public ResponseEntity<NicknameDuplicateCheckResponse> checkDuplicateNickname(@PathVariable("nickname") String nickname){
 
         userService.checkDuplicatedUserNickName(nickname);
-        return ResponseEntity.ok(new ResponseDuplicateCheckDTO(true, ResponseMessage.NICKNAME_AVAILABLE));
+        return ResponseEntity.ok(new NicknameDuplicateCheckResponse(true, ResponseMessage.NICKNAME_AVAILABLE));
     }
 
 }
