+++ src/main/java/com/WearWeather/wear/user/dto/request/RegisterUserRequest.java
@@ -1,4 +1,4 @@
-package com.WearWeather.wear.user.dto;
+package com.WearWeather.wear.user.dto.request;
 
 import com.WearWeather.wear.user.entity.User;
 import jakarta.validation.constraints.AssertTrue;
@@ -9,7 +9,7 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Getter
-public class RequestRegisterUserDTO {
+public class RegisterUserRequest {
 
     @NotBlank(message = "이메일은 필수입니다.")
     @Email(message = "올바르지 않은 이메일 형식입니다.")
