+++ src/main/java/com/WearWeather/wear/user/dto/request/VerifyEmailAuthCodeRequest.java
@@ -1,17 +1,22 @@
-package com.WearWeather.wear.user.dto;
+package com.WearWeather.wear.user.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(force = true)
 @AllArgsConstructor
 @Getter
 @EqualsAndHashCode
-public class RequestVerifyEmailDTO {
+public class VerifyEmailAuthCodeRequest {
 
     @NotBlank(message = "이메일은 필수입니다.")
     @Email(message = "올바르지 않은 이메일 형식입니다.")
     private final String email;
 
+    @NotBlank(message = "인증코드는 필수입니다.")
+    private final String code;
 }
