+++ src/main/java/com/WearWeather/wear/user/dto/response/NicknameDuplicateCheckResponse.java
@@ -1,11 +1,11 @@
-package com.WearWeather.wear.user.dto;
+package com.WearWeather.wear.user.dto.response;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Getter
-public class ResponseDuplicateCheckDTO {
+public class NicknameDuplicateCheckResponse {
 
     private final boolean isAvailable;
     private final String message;
