+++ src/main/java/com/WearWeather/wear/user/repository/UserRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    Optional<User> findByNickname(String nickname);
-    Optional<User> findByEmail(String email);
+    boolean existByNickName(String nickname);
+    boolean existByEmail(String email);
 
 }
